# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI-Powered School Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      summary: List all schools
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolList"
    post:
      operationId: createSchool
      tags: [schools]
      summary: Create a new school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInput"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /schools/{schoolId}:
    get:
      operationId: getSchool
      tags: [schools]
      summary: Get a school by ID
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    patch:
      operationId: updateSchool
      tags: [schools]
      summary: Update a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolUpdate"
      responses:
        "200":
          description: School updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"

  # STUDENTS
...
  /schools/{schoolId}/fees:
    get:
      operationId: listFeeRecords
      tags: [fees]
      summary: List fee records
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, overdue, partial]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeList"
    post:
      operationId: createFeeRecord
      tags: [fees]
      summary: Create a fee record
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeInput"
      responses:
        "201":
          description: Fee record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /schools/{schoolId}/fees/{feeId}/pay:
    post:
      operationId: recordFeePayment
      tags: [fees]
      summary: Record a fee payment
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: feeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /schools/{schoolId}/fees/summary:
    get:
      operationId: getFeeSummary
      tags: [fees]
      summary: Get fee collection summary
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Fee summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeSummary"

  # ANNOUNCEMENTS
...
  /schools/{schoolId}/dashboard:
    get:
      operationId: getSchoolDashboard
      tags: [dashboard]
      summary: Get school dashboard stats
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchoolDashboard"
...
  /schools/{schoolId}/dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
...
    School:
      type: object
      required: [id, name, code, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        principalName:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        studentCount:
          type: integer
        staffCount:
          type: integer
        subscriptionPlan:
          type: string
          enum: [basic, standard, premium, enterprise]
        createdAt:
          type: string
          format: date-time
    SchoolList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/School"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SchoolInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        email:
          type: string
        principalName:
          type: string
        subscriptionPlan:
          type: string
          enum: [basic, standard, premium, enterprise]
    SchoolUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        email:
          type: string
        principalName:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        subscriptionPlan:
          type: string
          enum: [basic, standard, premium, enterprise]

    # STUDENTS
...
    Class:
      type: object
      required: [id, name, schoolId, studentCount]
      properties:
        id:
          type: integer
        name:
          type: string
        section:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        classTeacherId:
          type: ["integer", "null"]
        classTeacherName:
          type: ["string", "null"]
        schoolId:
          type: integer
        studentCount:
          type: integer
        capacity:
          type: ["integer", "null"]
...
    AttendanceSummary:
      type: object
      required: [totalStudents, avgAttendanceRate, presentToday, absentToday, lateToday, byClass]
      properties:
        totalStudents:
          type: integer
        avgAttendanceRate:
          type: number
        presentToday:
          type: integer
        absentToday:
          type: integer
        lateToday:
          type: integer
        byClass:
          type: array
          items:
            type: object
            required: [classId, className, rate]
            properties:
              classId:
                type: integer
              className:
                type: string
              rate:
                type: number

    # FEES
    FeeRecord:
      type: object
      required: [id, studentId, amount, status, dueDate, schoolId, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        admissionNumber:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        feeType:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        discount:
          type: number
        status:
          type: string
          enum: [pending, paid, overdue, partial]
        dueDate:
          type: string
          format: date
        paidDate:
          type: ["string", "null"]
          format: date
        receiptNumber:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        schoolId:
          type: integer
        createdAt:
          type: string
          format: date-time
    FeeList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FeeRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    FeeInput:
      type: object
      required: [studentId, feeType, amount, dueDate]
      properties:
        studentId:
          type: integer
        feeType:
          type: string
        amount:
          type: number
        discount:
          type: number
        dueDate:
          type: string
          format: date
    PaymentInput:
      type: object
      required: [amount, paymentMethod]
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, online, cheque, upi, card]
        note:
          type: string
    FeeSummary:
      type: object
      required: [totalExpected, totalCollected, totalPending, totalOverdue, collectionRate]
      properties:
        totalExpected:
          type: number
        totalCollected:
          type: number
        totalPending:
          type: number
        totalOverdue:
          type: number
        collectionRate:
          type: number
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/FeeRecord"

    # ANNOUNCEMENTS
...
    SchoolDashboard:
      type: object
      required: [studentCount, staffCount, attendanceRate, feeCollectionRate, pendingFees, recentPayments, todayAttendance, classBreakdown]
      properties:
        studentCount:
          type: integer
        staffCount:
          type: integer
        attendanceRate:
          type: number
        feeCollectionRate:
          type: number
        pendingFees:
          type: number
        totalFeeExpected:
          type: number
        totalFeeCollected:
          type: number
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/FeeRecord"
        todayAttendance:
          $ref: "#/components/schemas/AttendanceSummary"
        classBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/Class"
        monthlyFeeChart:
          type: array
          items:
            type: object
            required: [month, collected, pending]
            properties:
              month:
                type: string
              collected:
                type: number
              pending:
                type: number
        upcomingBirthdays:
          type: array
          items:
            type: object
            required: [name, date, className]
            properties:
              name:
                type: string
              date:
                type: string
              className:
                type: string
...
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [admission, payment, attendance, announcement, staff_added]
        description:
          type: string
        actorName:
          type: ["string", "null"]
        targetName:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time